fixed errors and ran the program

+++ sample_code/integral.cc
@@ -1,10 +1,11 @@
 #include <ctime>
 #include <iostream>
+#include <math.h>
 #include "get_time.h"
 
 int main() {
     int segs = 10^8;
-    long long sum = 0;
+    long double sum = 0;
     timer t;
     for (int i = 0; i < segs; ++i) {
         sum += (sin(1 + i / segs) * log(1 + i / segs) + sin(1 + (i + 1) / segs) * log(1 + (i + 1) / segs)) / segs / 2;
